Add files via upload
Today is the Day 5 and I moved her Database out of the main code , So the DB in the Main code is been commented out, I only use it for reference now. 
I also corrected the set Alarm/Reminder date function . Initially, The function could not return a time or date greater than the current date or time.
Also, The secretary mode was added so as the scheduler mode. 

Right now, I am working on other Functions and Let see how this works out

Mustapha Ahmad